Now the data access caches the entire file on init
Before I was attempting to do a form of lazy loading, but the
performance was way below what I wanted it to be. Now it just
does a SELECT * on the Entry table, which is way faster when
the database is across the internet.

This is a fairly major change that fixes two performance issues:

  1.) Databases open MUCH quicker
  2.) Since file id's are now cached up front, the interface no
        longer hangs when you select an entry that contains a file.