Fixed a bug with the event of pressing escape
The event was being propagated out of the EntryView so the MainWindow was
handling it instead. That made it so it was impossible to clear the
sorting in the main window. Now the first time you press escape in the
entry view it clears the sort order, then the next time it lets the
MainWindow handle it which clears the view and search filters. So you just
have to keep pressing escape if you're focused on the entry view and you
want to clear the search filters.