No db connection needed to unlock application
Now if your application is locked and you lose connection to the database
(due to network drive error etc...) you can still unlock it and use it
read-only. You should be able to save-as, but any files in the database
won't be carried over because they are not in the cache.