Search widget uses dates instead of datetimes
I improved the interface and behavior of the search widget
by using a date edit instead of a datetime edit. Qt implemented
the spin-box style controls for me, so it was super simple. It
also uses a built-in calender popup for date selection, which is
much nicer than the arrows it was using before.