Improved the performance while closing a file
It is a really drastic improvement, bravo yourself! Actually it was
really low hanging fruit, because the original cleanup algorithm was
implemented clumsily. It used to delete each entry individually using
multiple transactions and updating the index and hierarchy as it went.

Now it does everything inside a single transaction, and it does everything
in O(N) time, rather than the horrible O(N^2) algorithm it was using.
It only uses 2 SELECT queries, rather than a non-deterministic number it was
using before.