Now the cleanup function does update the index
I decided to update the index anyways, even though we don't need to
right now, because everything is still fresh in my mind and I'm
afraid I'll have to do it anyways down the road so it would be easier
to do it now.

So now you're not required to call DeleteOrphans in the destructor,
but you at least have to wait until completion for the index to update.