Fixed a potential crash scenario with corrupt database
If you opened a database that has orphaned files, and those orphans happen
to be marked as favorites, it was crashing. Normally that never happens
because orphans are cleaned up when you close the file, but if cleanup
doesn't happen (crash, lost connection, etc...) you could render your file
unopenable. With this fix it ignores orphans on cache init, so it is much
more robust.