Implemented "Save As" much more efficiently
Now it very quickly creates a new database from the original inside
a single transaction. Also, it doesn't require an active connection
to the original database, since it uses the cache. The only gotcha is
that if you lose your database connection you will only be able to save
the files that are in the SQLITE driver's cache.