Fixed the behavior of the last keyfile location
Prior to this commit it would let an unprivileged user change the
preferences to remember the last keyfile, and it didn't prompt you again
for the location, it just remembered the one from before the preference
was changed. Now when you disable this feature, the last keyfile location
is really forgotten and you have to re-enter it again.